refactor: move to templates
